# backend/admin/docs/swagger.json
{
    "basePath": "/",
    "definitions": {
...
        "http.ArticleValResponse": {
            "additionalProperties": false,
            "example": {
                "val": "@article{key2024,\n  author = {Smith, J.},\n  title = {Title},\n  year = {2024}\n}"
            },
            "properties": {
                "val": {
                    "example": "@article{key2024,\n  author = {Smith, J.},\n  title = {Title},\n  year = {2024}\n}",
                    "type": "string"
                }
            },
            "type": "object"
        },
        "http.ErrorResponse": {
            "additionalProperties": false,
            "example": {
                "error": "error message"
            },
            "properties": {
                "error": {
                    "example": "error message",
                    "type": "string"
                }
            },
            "type": "object"
        },
        "http.TokenResponse": {
            "additionalProperties": false,
            "example": {
                "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            },
            "properties": {
                "token": {
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                    "type": "string"
                }
            },
            "type": "object"
        },
...
    },
    "info": {
        "contact": {},
        "description": "API for Furocoumarins admin backend",
        "title": "Furocoumarins Admin API",
        "version": "1.0"
    },
    "paths": {
        "/article/{id}": {
            "get": {
                "description": "Returns bibtex text for the given article ID",
                "parameters": [
                    {
                        "description": "Article ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.ArticleValResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                },
                "summary": "Get article by ID",
                "tags": [
                    "bibtex"
                ]
            }
        },
        "/auth/change-password": {
            "post": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "description": "Sends password reset link to user email",
                "parameters": [
                    {
                        "description": "Username or email",
                        "in": "formData",
                        "name": "uname_or_email",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                },
                "summary": "Request password change",
                "tags": [
                    "auth"
                ]
            }
        },
        "/auth/confirm-login-mail": {
            "post": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "description": "Exchanges word from email link for JWT token",
                "parameters": [
                    {
                        "description": "Token from email link",
                        "in": "formData",
                        "name": "word",
                        "required": true,
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.TokenResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                },
                "summary": "Confirm login via magic link",
                "tags": [
                    "auth"
                ]
            }
        },
        "/auth/confirm-password-change": {
            "post": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "description": "Sets new password using token from email link",
                "parameters": [
                    {
                        "description": "Token from email link",
                        "in": "formData",
                        "name": "word",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "description": "New password",
                        "in": "formData",
                        "name": "password",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                },
                "summary": "Confirm password change",
                "tags": [
                    "auth"
                ]
            }
        },
        "/auth/login": {
            "post": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "description": "Returns JWT token on success",
                "parameters": [
                    {
                        "description": "Username or email",
                        "in": "formData",
                        "name": "uname_or_email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "description": "Password",
                        "in": "formData",
                        "name": "password",
                        "required": true,
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.TokenResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                },
                "summary": "Login with username/email and password",
                "tags": [
                    "auth"
                ]
            }
        },
        "/auth/login-mail": {
            "post": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "description": "Sends a magic link to user email for passwordless login",
                "parameters": [
                    {
                        "description": "Username or email",
                        "in": "formData",
                        "name": "uname_or_email",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                },
                "summary": "Request login link by email",
                "tags": [
                    "auth"
                ]
            }
        },
        "/auth/renew-token": {
            "post": {
                "description": "Returns new JWT token for authenticated user",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.TokenResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Renew JWT token",
                "tags": [
                    "auth"
                ]
            }
        },
...
        "/pages/{name}": {
            "get": {
                "description": "Returns markdown content of a page from S3",
                "parameters": [
                    {
                        "description": "Page name",
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "type": "string"
                    }
                ],
                "produces": [
                    "text/markdown"
                ],
                "responses": {
                    "200": {
                        "description": "Markdown content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                },
                "summary": "Get page content by name",
                "tags": [
                    "pages"
                ]
            },
            "put": {
                "consumes": [
                    "application/octet-stream"
                ],
                "description": "Uploads markdown content for a page (admin only)",
                "parameters": [
                    {
                        "description": "Page name",
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "description": "Markdown content",
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Create or update page",
                "tags": [
                    "pages"
                ]
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "in": "header",
            "name": "Authorization",
            "type": "apiKey"
        }
    },
    "swagger": "2.0"
}